Delay line development and a lot of new hdl code
Decided I would start to check a lot of what I have been working into the repo.
Mostly because I am not sure the direction it is going and I am still learning
SV and want to be able to recall at least somewhat of what the process looked
like and keep something around that turned out being more complicated down the
road than it needs to be.

The SRL approach has seemed to be working so far (need to still verify it works)
but have able to test various configuration parameters for the module through
synthesis and the PE_ShiftReg at NUM=64, DEPTH=64, WIDTH=16 seems to be the
best.

However since then I have been tyring to:
A) figure out the best way to probe all the modules in a test bench to verify
    * this is looking like that it is going to be wasier to hard code hierarchy
    * paths for a few NUM instances of the module and then just verify din/douts
    * for when synthesizing large numebrs of these modules
B) and how to not have so many resets/en
    * while the PE_ShiftReg works NUM=64 still requires that there be 64 modules
    * each with a reset/en line feed to it which may not be required.
C) learning SV and some differnt appraoches to the language (interfaces/abstract
   classes/virtual methods and classes etc.)