example of connecting an interface to a dut
data_intf interface is an example of the canonical blackbox interface. It
doesn't have any other internal signals that it uses for whitebox verification.
It also has an example ofa modport and a simple monitor module below it.

Trying now to decide which is the best way to go to continue developing...