Show how to connect to probes with minimal tb code
Able to connect to the interface probes for whitebox verification with minimal
code that doesn't need to be hardcoded.

The bind syntax is able to target a specific instance or just all modules in a
design hierarchy. It is easier (and better) to just target all of them as it
removes some of the hierarchy paths needed in the tb for probing.

Once all the probes are bound the solution to then get a monitor to access all
of them is to use a generate block that creates NUM-1 initial blocks. There is a
question as to if this creates a race condition. However, I think we are ok
because this assignment is the only one in the initial block and when we start
driving signals it is after a reset that happens after time=0 of the simulation.

The only real open question now is to use virtual interfaces or the abstract
classes? Seemingly the downside to abstract classes in the interface is I am
wondering how we will get a unique ID attached for monitoring.