Testbench w/ working OSPFB monitor starting other verification
The OSPFB monitor now has a working handle to all shift registers in the delay
buffers and can print the status for each buffer in each PE. There is some
overhead in initializing all of that with the generate block prior to the main
initial block. There is also an open question as to why I need a packed array to
reference in the creation of the monitor classes. It has something to do with
the handle references (much like pointers in C) but I am not sure.

Starting some other verification features like looking at what built-in math
packages are available and other built in random number generators.