Improve source counter functionality and src_if interface
The source counter now dout is now a parameterized width with an additional
max_cnt parameter. This allows the counter to count more than just modulo on
max_cnt. Instead it can continue to generate processing and natural order
samples beyond max_cnt.

The src_if allows for a way to monitor the source and shows an example of how to
build a monitor and run task into the definition. The idea here was that this
class would up the modulo count from src_ctr to a real sample number for
verification based on a frame counter (number of modulo rollovers). But then I
realized this information wouldn't be possible to keep connected to the sample
and would not be able to generate a correct sample number everywhere. This is
what led to improving the source counter capability.

However, this src_if may not be very useful now. But am keeping it around as an
example if it is needed elsewhere. Because the idea is similar to what I would
want to do to observer the samples/capture data and then have something produce
a "golden" value indpendent of the module (like what was done with the python
verification engine).