Add natural ordering to src_ctr, remove src_if in current implementation
The src_ctr now supports natural order counting. I don't have it documented
anywhere but to my best recollection the processing order counter was first
used in the python code to verify the phase compensator. Then natural order has
always been what is the input into the PFB. Then they were combined into a
single python class so that there was a single source that could be used to
check both in different situations.

But processing order for the phase comp buffer because this order matches the
ordering coming out of the ospfb. Then natural matches what is on the input.

The src_if was also removed. But the skelton of what was written is left behind
as an example for how to use tasks in an interface class since that was where
that interface was headed.