ospfb fir works, tb reads file for verification, add valid buffer to ospfb monitor
PE module had dout on the output instead of sout

To include the valid buffer in the ospfb monitor a typedef for the sr_probe
doesn't work well anymore since the valid buffer width is one bit wide. So
instead of creating 3 new typedefs the old ones have been removed and a WIDTH
parameter added to the OSPFB monitor class that will create handle of the
correct width.

The tb is then updated to now use the valid buffer and connect the handles to
the HDL instance.

The ospfb polyphase fir verified to be working by comparing with python model
output. There is a problem with the idea for how the get_mac() method works for
the PE monitor. The issue is that the get_mac creates the mac at the instance of
the data coming into the PE instead of coming out. The get_mac method needs to
be removed or fixed.

Since the python simulator works and is compatible with different data types we
can capture int16 samples into the python model and compare the output with the
hardware simulator. It is shown that the values line up and that python could
potentially be used to debug some of the scaling and rounding work to be done.