substantial rework of python simulator
A new apps directory for placing scripts using the ospfb and other utility
classes for producing test cases and verification data

The main in the ospfb now shows mainly examples for how to make instances of the
different classes and interface with them.

ospfb now accepts the taps as a parameter with utility classes modeled after the
Source classes for generating filter taps.