seperate combined zoom spec and symbolic verification into seperate scripts
For readability and reusability the zoom spectrometer with complex and the
symbolic verification are now seperate apps.

The ospfb is now slimmed down to the essentials continuing to show how to
initialize some of the libraries and components

the initval of the ospfb needs to be revisited if arbitrary valid verification
of symbolic inputs wants to be revisitied. It used to work but overtime has been
deprecated.

This happened when the ospfb stopped producing samples after the first proof
that this version was working. When the ospfb then started to make use of
different data types the class itself stopped producing data and initval lost a
lot of its support.

Notes are still around though for what needs to be changed to support these
arbitrary values