standardize language for 'order' param on sink/source classes
"natural" and "processing" order are the terms chosen and refer primarily
to how samples are read at the output of the polyphase fir and phase
compensator. However, they also affect how samples are generated (either natrual
or processing order).

This commit clarifies that the ability to specify an order was primarily chosen
to verify the polyphase fir and phase compensator independently. Verifying both
components together only natural order makes sense because this is how samples
are practically delivered to the core.