Implement ram_if interface and probe for phasecomp
This interface and probe looks a lot like the other verification
probes/interfaces. But also is included a lot more functionality because it was
anticipated to need those to debug some of the phasecomp issues lining up with
the FIR resampler. Those didn't end up getting used but shows how you would
extend the current probes.

Also shown is comments and ideas for how to generalize the probe interface to
something more generic that it could work for the ram and srl's and others. But
it wasn't done and there may be issues finishing that. There were some problems
with getting packed/unpacked types to co-operate and so getting something to
truly work would require type templated classes and the like