Working integrated SV `PhaseComp` into `OSPFB`, correct python and sv phasecomp
Because of the hardcoded "enable after the first (en=2)" cycles on the phase
compensation buffer in the python simulator version I anticipated there would be
issues wiring the two together. Uncovered a few inconsistencies including why
the magic hardcoded value.

First, the original "enable after" was after M*(P-1)+2 cycles. But it turned out
that it was any multiple of M cycles. Which intuitively this made sense since
the phase compensation buffer had to be running at the upsampled rate and
outputting a valid sample every cycle. It made sense that it would work
correctly modulo M.

This brought it down just `2` which is meant to enable the phasecomp on the
second cycle. Which didn't make sense. Because the python code was just not
running the phasecomp on that first cycle. Which is when it dawned on me that
the phase comp just needed to be held until it aligned with the polyphase fir
output meaning that the initialized decimator (commutator) phase was not
correct.

This made perfect sense because the polyphase FIR resampler starts injecting
samples (on an equivalent parallel device -- but the systolic version makes it
all serial) at port 0 and then rolls over to port M-1 (really D-1 in the
equivalent parallel device but the tready/tvalid goes down for the first M-D
until commutator aligns at D-1).

The polyphase FIR has this initial phase because in the symbolic, canonical
equation this is where the first causal sample is inserted into the PFB. Since
the polyphase FIR was starting at port zero the phase compensation buffer needs
to do the same. (The phase compensators job is to pull samples off the polyphase
FIR in processing order -- from port M-1 to zero -- and produce them in natrual
order -- from port 0 to M-1 -- since the FFT expects them this way)

This commit therefore includes this capability in both the python and
system verilog versions. With this implemented there is no need for an "enable
after" signal and the phase comp can run non-stop. In python there is still an
enable after to support symbolic simulation results, but that should be looked
at because I am pretty sure it can be removed with improved checking and
handling of the symbolic data case.

Second, the PhaseComp module had implemented the wrong count direction on the
modinc (it was subtracting modinc instead of adding) where subtracting is only
useful when verifying operation.

Third, the question of asynch vs synchronous outputs came up again. I had run
into this decision early in the SRL development where my outputs where "off by
one" (one cycle later than I wanted) because they were being synchronously read
out. Since I had chosen to make the SRLs asynchronous earlier I changed the
PhaseComp module to also do asynchronous for now.

However, for timing it would be better to have synchronous outputs (this is the
recommendation by Xilinx and makes sense -- you are adding an extra cycle so you
don't have the timing required to pull bits out of the RAM). Changing to
synchronous in the future shouldn't be too hard as long as continue to not
accept back pressure.

Also, added both golden outputs for polyphase fir and phase rotation output.
Keeping the polyphase fir outputs would be helpful if a module wanted to assert
and probe those intermediate values.