Integrate Xilinx FFT into OSPFB Module
The Xilinx FFT has been integrated into the OSPFB module with additional
modifications to make this fully compatible. The FFT operation has not been
tested yet. Just that the addition of the XFFT and changes described by the
following have not broken what has been previously shown to work.

Since the XFFT is a complex FFT the OSPFB module needed to be adjusted so that
it would process the complex data stream (I and Q and there fore the slv axis
interface now accepts real and imaginary data similar to the XFFT
specification of {Q, I} ordering.

Obviously with the addition of the second data stream an additional FIR data
path was added.

Incorporating the XFFT running in real-time streaming pipelined I/O mode
introduced some modification to the start up logic. While we are indifferent to
when we start capturing data on the output the OSPFB must move in lock step,
meaning that the phase compensation output needs to be rotate at the same frame
timing of the FFT and so to achieve this the FIR front end, phase compensation
and FFT must coordinate coming on line.

To do this the state machine that is driving the slv axis interface of the ospfb
indicating that it is ready to receive samples (FORWARD) or that it was
resampling (FEEDBACK) an additional state was added to wait for the FFT
(WAITFFT) by asserting tvalid on its interface once coming out of global reset
and that an internal `hold_rst` signal would stay high until the FFT comes out
of its startup wait state and indicate it was ready. As it is ready `hold_rst`
is de-asserted and the FIR front end and phase compensation buffer starts
operating at the initial SRT_PHA=DEC_FAC-1 position (port 0).

To test the integration of the FFT and the new additional startup state machine
the test bench was modified to still capture the phase rotated fir data and
check that the output still matches as expected.

The `wave_wait.do` is the signal questasim do file used to show the startup of
the state machine and transition of states and that the tready from the fft is
sensed at the right time.

The next step is to add real data on the input. But it was realized that we
first need to work on the dual clock fifo and cdc.