Working Hardware Simulation of Full OSPFB
This is the first complete hardware version of the OSPFB.

The work from learning the dual clock fifo has been included to create two
versions of a top model. The first is the simple counter and the other uses the
adc model for trying to show spectrum with simulated complex tones.

Note that there are some limitations when working with the ADC model that prove
as examples for additional work that needs to be done be fore this is truly
complete. But the configuration in this commit will produce the same output data
file that shows the output tone when plotted.

Specifically, the ADC models scales the tone to one-tenth of the FS input. The
coefficients are equal value (a rectangualr window) such that the sum of the
weights equal approxamitely equal the full-scale coefficient width. In this case
it was 12-bit coefficients and it came out that each tap was '0x000a' (decimal
10). Also the output spectrum is reversed, meaning that the output frequency is at
1-f_normalized where you would have expected it in a matlab simulation. This is
something to also be worked out.

Without the ADC and coeffecients set as they are the spectrum is otherwise noise
and the fact that as you scale eveything down seems to point that we do have
overflow issues in the interstages and that scaling and correcting this
computation just became higher priority (as I expected it would).

With the two top modules and testbenchs the counter source is still viable to be
eligable to check outputs with the python version. However, since scaling has
become more important and we have seen that python seems to behave with the same
overflow behavior we may be able to implement the scaling in python to check and
that with a python version of the ADC source I may be able to have python
version to compare and check against the hardware model.

Fixed a bug where the start up state machine was going to FORWARD state first.
The reason this was a bug is that because that the initial phase step that
everything is starting on is really the last count of the forward state and so
the right thing to do would be to go to FEEDBACK.

Also a bug where the s_axis data first sample was being read out of the fifo but
not set to din of the PEs. Things like this could kill me later if I am not
careful.