Bring counter with cyclic ramp coefficients for fir output verification
The last commit with the ADC removed initializing filter coeffecients in the PEs
and instead read in coeff files. The capability to test the polyphase fir
outputs with the python simulator has been using cyclic ramp coeffecients (i.e.,
coeff that are the same for each PE that count from 0 to M-1).

Added coeff files to again be able to test the counter mode and continue to use
python simulation model.