Add fifo between phase comp buffer and real-time xilinx fft
The Xilinx FFT in real-time mode has a bug. The documentation had said that in
real-time mode one word is buffered prior to starting frame processing (pg. 52
verison 9.2 June 17, 2020). However, it is actually two words that are passed to
the core before processing. This was shown using the impulse test sequence to
look for when an impulse is present as the first sample to the fft in relation
(providing a constant output) and the event frame started signal was high.

This meant that the timing of the core with the polyphase fir and phase comp out
of initialization was off and would need to be re-done. Previously, since the
documentation mentioned that only one word and that the core de-asserted tready
upong seeing tvalid for the first time, took a sample, and then asserted event
frame started signal I took the tready going low as an OK time to hold the
circuit in reset. However, this is not true because apparently the core
transfers that sample when tvalid is asserted, goes ready low, then buffers a
second sample and starts processing the following cycle (hence the 2 instead of
1 sample buffer).

Because the polyphase FIR cannot accept back pressure and there is no real good
way to get any other timing out of the core the only option seemed to insert a
small XPM AXIS fifo between the phase comp and the FFT. After doing this I was
able to assert the impulse into the FFT with the same expected behavior as my
python simulation showing that the hardware matched.

The problem with the FIFO is that it is a small 16 deep fifo when only a 2 deep
fifo is needed. I started coding a fifo in ringbuffer.sv but gave myself a time
limit to do so and didnt meet that. So I left it with the XPM fifo until needing
to change. The idea is that hopefully a distributed RAM FIFO will take the same
resources (or close to) a two deep version.

Adding the delay fifo removed the wait FFT state. However, if there are
additional wait states there is nothing to detect them and help the OSPFB catch
up in any meaningful way which could present issues.

Right now the real and imaginary streams are concatenated at the input of the
fifo. While that works in simulation I don't know what the cost is in synthesis
and so we may need to pipeline that out.