Verified ospfb functionality
Although previously committing that the OSPFB was working there were still open
questions if the architecture checks out.

The impulse sequence is one of the best inputs (in my opinion) and it showed
that the timing for of the polyphase fir was off with the Xilinx FFT because of
documentation discrepencies. Then after this quantization was still a question
because of the shallow roll-off and notch in the spectrum.

Python was shown during intial development that np.int roll over like hardware
and so continued to use the ospfb and poke it with int16 and int32 data and
coefficients to see how much I can get the hardware to match python (knowing
that there are limitations since the fft is floating point).

In the python simulation I observed that using int32 and scaling the
coefficients down until no overflow (9 bits) but keeping 12 bit data, M=64, P=4
I observed similar spectrum but would eventually become the steep deep spectrum
I expected. So I was able to see that the shape was expected due to the filter
wind up but because of the 32 bits and the floating point fft we were able to
get the desired spectrum.

This pointed to that indeed the arithmetic is wrong and getting the correct
slice and tunning the hardware for the right signal conditions will be
important. After checking that the system verilog adc model output was similar
to the python rfdc and tone generator I then tweaked the fixed point python bit
values to fit within int16's without overflow since this was the width used in
the hardware. With ADC_BITS=8, COEFF_WID=4, M=64, D=48, P=4 the python simulator
showed that the spectrum was still acceptable at 4 bit coeff widiths and 8 bit
data. There was just slight quantization in the spectrum noise floor as
expected.

Tweaking the hardware simulator the concerning artifacts in the spectrum were
resolved. It was nice to be able to adjust the hardware so quickly by having
parameterized the hardware.

From here there is a lot to do but could work on optimzation since correcting
the arithmetic and slice ability are needed but terribly break things.

todos:
  * continue to improve fixed point capability in python
  * correct rounding/slicing arhitmetic (perhaps custom scaling schedule)
  * take dumped data from hardware do a second stage process (this will really
    be one of the last things to prove it works)
  * optimzations (parallel samples)
  * add axi memory mapped fifo instead of axis so easier transition to hardware
  * start running through synthesis