Fixed bug found when scaling to M=2048
State machine bug from adding the delay fifo caused timing to be off. The `vin`
assertion was not being asserted on the first data sample coming out of the
wait fifo state. Also was being set twice and not default value of zero at the
top of the state machine logic.

Tested values of M=64, 256, 512, 2048 with counter, impulse and full windowed
coefficients everythong seems to check out.