Simplified top module that can be handed to vivado synthesis
The restructuring has now allowed for easily handing hardware to vivado and
keeping simulation versions. This is a top file we can use to get synhtesis and
reports from vivado. It has ran all the way through route (although with out
xdcs)

But the utilization report shows us that we are using about 25 percent of our
LUTRAM resources.... which isn't good. But we expected this so it is on to
optimizations