Remove m_axis_fir port from ospfb and simplify adc/impulse tb
The m_axis_fir port was used for getting direct access to the the output of the
polyphase fir + phasecomp output of the ospfb. However it adds complexity and an
unecessary step to remove the port to pass to vivado synthesis. Since we won't
want it any way we need to remove it.

The counter tb now grabs a direct reference the output of the fir for
comparison.

The adc/impulse tb have removed the unnecessary reading from a golden file and
simplified to just waiting for the dummy axis vip to be filled.