working on xpm vivado synthesis
The synth_xpm_ospfb can make it through vivado synthesis but only when I remove
the parameter for the generate block that provides the filter taps. Vivado
synthesis cannot work with string variables (only parameters that are not system
verilog string typed) and so doing a string interplation to fill in the generate
block for the correct coeff file doesn't work.

I really can't figure out a way to do it with any sort of built in language
support... I really will have to hard code PTAP instance PE's to pass the 8
files as parameters. Or have all the taps of the filter hard coded in a
package and then passed in as a parameter.

started to add missing connections to xpm primitives so that synthesis provides
less warnings and that it may better remove unused ports and improve the design