Adding the playground code used to figure out getting taps to different pe rams
These codes were the ones used to help figure out how to get the taps to
different PE RAMs and get through Vivado synthesis.

Vivado synthesis doesn't accept string parameters and there is no way to do
parameterized strings with like $sformats or $psprintf.

I had thought about a lot of a few ways to get the taps in the PE:
  * One RAM at the upper OSPFB level that then loads each PE as part of an init
    sequence. This seemed valid but seemed to imply extra memory overhead.
  * Have seperate files and hardcoded PE instantiations. Seemed like a lot of
    overhead and hard to maintain. Seemed like that was a terrible idea and I
    can't believe that is how it is done in real system.
  * This one is along the lines of what I did but still want to know if it can
    be done. Use systemverilog struct/typedef/enums to get the taps present. But then I
    don't know how to reference specific elements of the struct

The best thing I could do was to have a python script create a python package
that would then be imported and then past as an instantiation parameter to the
OSPFB and each PE in the generate block gets a range of those taps.

This fir code isn't tested. But it seems like it could work with more effort. It
is based more on an AXIS protocol. But the taps certianly do get in and I have
passed it through synthesis and see them show up.