Some SystemVerilog typing expoloration and implementation notes
In starting the parallel input fft and necessary complex multiply (for twiddle
factor recombination) I had a lot of questions related to my frustration with
getting system verilog to go through vivado synthesis and also how I was going
about structuring the parallel fft in terms of the two sample per clock word,
the arithmetic, etc.

These files are the result of that investigation. The readme highlights where I
was able to get with this and summarize some of my thoughts and conclusions at
the time regarding using interfaces, wrappers for interfaces, and typing.

For the parallel fft I ended up doing something similar to what is in cmpx_v1_5
(i think... idk... check the readme and parllel fft to be sure).