`sample_delay` module
To align the input samples correctly with the decimation phase of the
oversampled pfb a single sample delay needs to be inserted. This is because the
data coming from the rfdc are sequenced {x1, x0}, {x3, x2}, {x5, x4}, ... but
the first decimated time output of the ospfb relies only on time sample x0. So
the first time samples {x1, x0} split two different outputs. The solution is
this module that inserts a single sample delay to align the filterbank
correctly.