Update `xpm_delaybuf` to use `alpaca_data_pkt_axis`
The delaybuf and its testbench now are compatible to take multiple (or single)
samples per clock based on the `alpaca_data_pkt_axis` interface the module
receives.

No internal logic of the delaybuf needed to be changed. This is because in the
systolic array formulation the number of delays is corresponding to the word and
so we just always count by the depth of the FIFO.