use interfaces for fixed-point representation, create cx mult-add module
Modifed the module ports to use interfaces. The interfaces incorporate the
ability to generalize for fixed point data. It also includes a method to convert
the bits to display the fixed point value in testbenches for comparison.

This module can not be used in the alpaca butterfly so while it solved the FIR
PE multiply add problem it didn't address the problem that got me here in the
first place. Which is to correctly scale the addition for the alpaca parallel
fft operation and to also do it efficiently in hardware. (To fix it as it stands
would have put an arithmetic shift in alpaca_butterfly where `Xkhi/loi` where
created in an `always comb` block but I wasn't certain this would perform well
for high performance.) So considering this it seemed more relevant to create the
complex version of a multiply-add that uses the same fixed-point conveience
constructs.