check with vivado synthesis in non-project mode
prepared for a vivado synthesis check by making a simple top module that
exposes and connects interfaces. Removes the testbenches and puts them in
seperate source files.

`alpaca_multadd` module checks out with vivado synthesis. Everything is inferred
as I would hope.

`alpaca_cx_multadd` however I have a question why the `c_re` and `c_im` are
being but into SRL Shift registers. I am not sure if the `cr_delay` and
`ci_delay` these c registers are beign absorbed to create the enitre delay
(because that is what I want) or if the behavior is completly wrong and this
will mess up perfomance and speed.