cmpx multadd to include difference for parallel xfft, modify structure for merge
Needed to add the multiply subtract output so that it better matches what the
core is needed for in the parallel xfft module by providing both frequency bins
on the output in the same clock cycle.

Checked with synthesis there is still a confusing output about the C delay
registers being an SRL and I don't see them being absorbed into the DSP and so
the extent that this will become a problem I expect to see when testing max
clock frequency.

Better made the testbench and typing consistent to help with integration in the
baseline