Incorporate processing of multiple samples per clock
This brings the OSPFB back to where it was when the first version of the core
was completed but now able to process multiple samples per clock.

The polyphase fir has only been tested with two samples per clock but everything
should be in place to handle any even number of samples per clock (e.g., 2, 4,
8) as long as it is even. The oversampled parameter of `M-D` must also be
greater than 16.

Only two samples per clock was tested because that is effectively the limit of
this core since the parallel fft was done specifically to only handle two
samples per clock. Another core such as the CASPER or ASTRON FFT would need to
be used or additional work would need to be done to do the arbitrary radix
stage to recombine the twiddle factors for the parallel input fft.

This also brings in the pipelined complex multiply-add/sub needed in the alpaca
butterfly module.

Also included are the other supporting infrastructure and bringing the
testbenches up to date for the support. The python ospfb fir was tested with the
hdl version by adding a modulo rollover counter so that the impulse generator
coudl instead be used instead of an up counter. It wasn't obvious at the time
how to do it. But I think something similar to the ADC may be possible if I
wanted to go back and implement that. But for now the change in the impulse
counter would need to be made to test the ospfb python simulator with the
counter to verify the polyphase fir.

From here is to incorporate the multiply-add module into the pipeline and see
what changes need to be made to maintain the correct delay between pe's.