Addition of axis phasecomp module, update to this module in datapath
The old phasecomp would be held until `hold_rst` was released by the dual-clock
fifo `tvalid` signal at the input of the ospfb. When valid data started the
phasecomp would come online in tandem with the rest of the core (starting at the
same decimated phase -- which is processing starting at port zero when thinking
about the device as a parallel port device -- or in other words the last step
before updating the phase offset in terms of the phasecomp).

Not being axis compliant seemed to be a problem as I started to add a fully
pipelined multiply-add in the PEs and so went ahead and made it axis to be able
to test the new multiply-add.

In the process of debugging the core, when comparing simulated outputs using the
new core and the old core they did not align perfecly until I removed the extra
`vldd` dealy register in the phasecome on the axis `tvalid`. The `vldd` second
delay register on the valid signal was added because it was needed to align the
ouputs correctly when looking at the output of a testbench just looking at
phasecomp outputs.

So the question is if this bug has been here the whole time or if I just
squashed a bug that would have presented itself later when say looking at zoom
mode second stage processing outputs. I have opted for now to go with leaving
that register there because when comparing with the old outputs they are flatter
in the noise floor (the old outputs had a slight ripple to it).