Add pipelined multiply into PE
Pipelined multiply in the PE results in the same simulation outputs with
previous multiply-add that was being done in a single clock cycle. To do this
the added latency needed to be added on the `axis_sumbuf.tuser` that represents
the `vin` signal for the PE that lines up all the computations. Also needed to
add a delay between the `loopbuf` (resampler) and `databuf`. Since the
`alpaca_multadd` adds `MULT_LAT` cycles buffers needed to be added on equivalent
signals across the PE.

Due to the phasecomp and now multiply-add module adding pipelined latencies the
python polyphase FIR golden counter model testbench no longer lines up correctly
with the hardware outupts as the python model doesn't yet include these
pipelined registers. Instead of going back and adding them (which needs to be
done) I just added additional zeros to line up the outputs.

The most useful thing to evaluate it working was to continue to use the old
multiply and add the latencies where needed to make sure the added delays could be
validated independently of the proper multiplcation of the new `alpaca_multadd`
module.

Note, that while the phasecomp and PEs use axis to begin processing
they aren't fully axis comliant as the tready doesn't fully support back
pressure use of the tready (to stop or report an error) and tlast can also be
added.

This is just a snapshot after getting it to work, going to go back now and
cleanup the code.

There were a few bugs in the `alpaca_multadd` that needed to be added from when
working on the adding `alpaca_cx_multadd` from the parallel xfft work. Also
duplicated use of `a_delay` that would reuse `a` for the `b` value. I would have
been able to spot this sooner had I not rushed doing the testbench for the core
by starting evetythin at `1` and counting up.

Next, need to investigate why using the `hann` window filter coefficients with
`COEFF_FRAC_WID` also equal to zero when it should be `COEFF_WID-1` like with
the twiddle factors. It probably means I am not exporting them correctly.