Revisiting the core ospfb/cs-pfb/bank of lpf algorithms in python
I am currently determining if the hardware modules are working correctly.
There are a few processing artifacts that we haven't encountered

e.g., most of these are when noise is absent from the signal - the pruned second
stage spectrum has what seems to be alias tones, pruning the second staage
outputs can result in slight scalloping being present, there does not seem to be
one formula to follow to determine the channels that are overlapping and need to
be discarded, the bank of lpf show a "pedastal" structure after a second stage
fft, just using an FFT in the second stage results in spectral leakage showing
bad lobe artifacts.