Improve adc model with awgn, xfft using convergent round even
Reduced FFT size to evaluate and finish out the remaning work on the ospfb after
finishing the multiple sample per clock. The multiple sample per clock was
developed at M=512 since the minimum axis fifo depth needs to be greater than 16
and I was wanting to be ambitious and also tackle adding pipelined multiply-add
in the PE and was unsure if a solution was going to be able to present itself
that allowed to borrow the required clock cycles from the fifo buffers. This is
not the case however because I forgot the fifos require a power of 2 depth and
so borrowing latency from them won't work.

The adc model now MSB aligns the output data and includes an AWGN generator
using system verilog's random stability capabilities with $dist_normal, $urandom
and seeding independent processes.

The XFFT template has also now been configured in Vivado to use convergent
rounding to even on the output samples. This has made an incredible difference
in the quality of the output spectrum. Prior to this there were several very
concerning artifacts (a raised floor near zero in the coarse outputs, spikes in
fine spectral outputs where that raised floor was, generally just really crumby
noise floor). By far the biggest improvement has come from convergent rounding
in the fft and was easy to do. After adding the convergent rounding the only
real artifacts were the DC bias from truncating in the PE.

Other modifications to support reducing M (scaling schedule, config widths,
etc.)